Add files via upload
Front-End:
We can create the front-end of the to-do list application using HTML, CSS, and JavaScript. Here's a HTML structure for the to-do list.
In this HTML file, we have a title, an input field for adding tasks, a button to add tasks, and an unordered list ('<ul>') to display the tasks.